add nodemon / add crypto.js
