+++ cleanup.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-rm -rf wallet/
 rm -rf node_modules/
+rm -rf wallet/
 rm -rf db/
 
 rm nohup.out
