+++ src/modules.js
@@ -1,4 +1,5 @@
 export * from "./utils";
+export * from "./crypto";
 export * from "./database";
 export * from "./types";
 export * from "./blockchain";
